+++ CHANGELOG.rst
@@ -48,6 +48,7 @@ Added
     - ``SHEETS_SYNC_ENABLE`` Django setting (#2247)
     - Periodic celery task setup logging (#2247)
     - Disabled iRODS data request batch selection checkbox tooltips (#2295)
+    - ``perform_project_sync()`` iRODS ticket updating (#2133)
 - **Taskflowbackend**
     - Project viewer role support (#2200)
     - ``user`` arg for ``TaskflowAPI.submit()`` (#2157)